Add new class AliensList.java, edit class Alien.java and code refactor
